added unet

+++ main.py
@@ -426,8 +426,8 @@ def main():
             testing = subjects[i * n_test:(i + 1) * n_test]
             not_testing = subjects[(i + 1) * n_test:] + subjects[0:i * n_test]
             if val_split > 0:
-                validation = not_testing[:len(not_testing) * val_split]
-                training = not_testing[len(not_testing) * val_split:]
+                validation = not_testing[:int(len(not_testing) * val_split)]
+                training = not_testing[int(len(not_testing) * val_split):]
             else:
                 training = validation = not_testing
